+++ src/cli/progress.ts
@@ -112,23 +112,3 @@ export class BulkOperationProgress {
     return this.failures.length > 0;
   }
 }
-
-/**
- * Progress indicator for deployment pipeline
- */
-export class ProgressIndicator {
-  private reporter = new OraProgressReporter();
-
-  startSpinner(text: string): () => void {
-    this.reporter.start(text);
-    return () => this.reporter.succeed();
-  }
-
-  complete(text: string): void {
-    this.reporter.succeed(text);
-  }
-
-  fail(text: string): void {
-    this.reporter.fail(text);
-  }
-}
