+++ src/core/deployment/index.ts
@@ -2,6 +2,7 @@ export { DeploymentPipeline } from "./pipeline";
 export { MetricsCollector } from "./metrics";
 export { DeploymentSummaryReport } from "./summary";
 export { DeploymentReportGenerator } from "./report";
+export { ProgressIndicator } from "./progress";
 export { getAllStages } from "./stages";
 export type { DeploymentContext, DeploymentStage, DeploymentMetrics, EntityCount } from "./types";
 export type { DeploymentReport } from "./report";
